Real raw eval
Bench: 4085577
